refactor:moved toGroupLabel method to pipe folder
